requirements.txt
pdfkit==1.0.0
tabulate==0.9.0
pytest-mock==3.14.0
